Fix '4_ocp_update.yaml' playbook when setting of external IP is enabled
It didn't check for existing value and when the existing value is the
same as the one for patch then it fails with "not patched" error.
So, fix it by adding check for the existing value and skip patching if
the value is already the same.

Also, remove duplicated task from the same playbook.